+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Section Order:
 - `tasks.open_fleet` refactored
 - `tasks.send_fleet_invitation` refactored
 - `tasks.check_fleet_adverts` refactored
+- `tasks.get_fleet_composition` refactored
 
 ## [2.6.0] - 2025-08-14